(20220309-025700) 1. Nightly check in, Completed first 40 Enum Test cases. One compilation error error, will fix tmorrow.

+++ staging/src/main/java/com/hwtsllc/fixengine2022/fix27/tags/Tag14CumQty.java
@@ -28,7 +28,7 @@ public class Tag14CumQty extends TagTypeAbstract {
 
     private QtyType qtyType;
 
-    public Tag14CumQty(long qtyType) {
+    public Tag14CumQty(int qtyType) {
         setFixType(FIXType.FIX14_CUM_QTY);
         setDataValue(String.valueOf(qtyType));
         this.qtyType = new QtyType(qtyType);
@@ -39,7 +39,7 @@ public class Tag14CumQty extends TagTypeAbstract {
      * @param args      Not used.
      */
     public static void main(String[] args) {
-        Tag14CumQty tag14CumQty = new Tag14CumQty(1234L);
+        Tag14CumQty tag14CumQty = new Tag14CumQty(1234);
         System.out.println(tag14CumQty);
         System.out.println(tag14CumQty.toLogStringVerbose());
         System.out.println(tag14CumQty.toFixTagValuePairString());
